Move commands tests to designated location
