+++ src/core/tests/command/test_command.py
@@ -5,35 +5,8 @@ import toga
 import toga_dummy
 from tests.utils import order_test
 
-from toga import GROUP_BREAK, SECTION_BREAK
+from tests.command.constants import PARENT_GROUP1, COMMANDS_IN_ORDER
 from toga_dummy.utils import TestCase
-
-PARENT_GROUP1 = toga.Group("P", 1)
-CHILD_GROUP1 = toga.Group("C", order=2, parent=PARENT_GROUP1)
-CHILD_GROUP2 = toga.Group("B", order=4, parent=PARENT_GROUP1)
-PARENT_GROUP2 = toga.Group("O", 2)
-CHILD_GROUP3 = toga.Group("A", 2, parent=PARENT_GROUP2)
-
-A = toga.Command(None, "A", group=PARENT_GROUP2, order=1)
-S = toga.Command(None, "S", group=PARENT_GROUP1, order=5)
-T = toga.Command(None, "T", group=CHILD_GROUP2, order=2)
-U = toga.Command(None, "U", group=CHILD_GROUP2, order=1)
-V = toga.Command(None, "V", group=PARENT_GROUP1, order=3)
-B = toga.Command(None, "B", group=CHILD_GROUP1, section=2, order=1)
-W = toga.Command(None, "W", group=CHILD_GROUP1, order=4)
-X = toga.Command(None, "X", group=CHILD_GROUP1, order=2)
-Y = toga.Command(None, "Y", group=CHILD_GROUP1, order=1)
-Z = toga.Command(None, "Z", group=PARENT_GROUP1, order=1)
-
-COMMANDS_IN_ORDER = [Z, Y, X, W, B, V, U, T, S, A]
-COMMANDS_IN_SET = [
-    Z, GROUP_BREAK,
-    Y, X, W, SECTION_BREAK, B, GROUP_BREAK,
-    V, GROUP_BREAK,
-    U, T, GROUP_BREAK,
-    S, GROUP_BREAK,
-    A,
-]
 
 
 class TestCommand(TestCase):
@@ -129,52 +102,3 @@ class TestCommand(TestCase):
     )
     test_order_commands_by_groups = order_test(*COMMANDS_IN_ORDER)
 
-
-class TestCommandSet(unittest.TestCase):
-    changed = False
-
-    def _changed(self):
-        self.changed = True
-
-    def test_cmdset_init(self):
-        test_widget = toga.Widget(factory=toga_dummy.factory)
-        cs = toga.CommandSet(test_widget)
-        self.assertEqual(cs._commands, set())
-        self.assertEqual(cs.on_change, None)
-
-    def test_cmdset_add(self):
-        self.changed = False
-        test_widget = toga.Widget(factory=toga_dummy.factory)
-        cs = toga.CommandSet(
-            factory=toga_dummy.factory,
-            widget=test_widget,
-            on_change=self._changed
-        )
-        grp = toga.Group('Test group', order=10)
-        cmd = toga.Command(
-            lambda x: print('Hello World'),
-            label='test',
-            tooltip='test command',
-            shortcut='t',
-            icon='icons/none.png',
-            group=grp,
-            section=1,
-            order=1,
-            factory=toga_dummy.factory
-        )
-        cs.add(cmd)
-
-        self.assertTrue(self.changed)
-        self.assertIsNotNone(cmd._impl)
-
-    def test_cmdset_iter_in_order(self):
-
-        test_widget = toga.Widget(factory=toga_dummy.factory)
-        cs = toga.CommandSet(
-            factory=toga_dummy.factory,
-            widget=test_widget
-        )
-        commands = list(COMMANDS_IN_ORDER)
-        random.shuffle(commands)
-        cs.add(*commands)
-        self.assertEqual(list(cs), COMMANDS_IN_SET)
